+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # baasix
 
+## 0.1.11
+
+### Patch Changes
+
+- Added tenant scoping as setting in schema definition itself.
+- Updated dependencies
+  - @baasix/types@1.0.6
+
 ## 0.1.10
 
 ### Patch Changes